opdatede models

+++ Vitec/Models/Card.cs
@@ -7,7 +7,7 @@ namespace Vitec.Models
 {
     public class Card
     {
-        public int cardID { get; set; }
+        public long cardID { get; set; }
         public int SecurityNumbers { get; set; }
         public DateTime ExpDate { get; set; }
     }
